architecture: move chess AI timer from module-level _timer to useChessAI hook
- Extract buildRecord/rateMove/toNotation to store/chessRecordUtils.ts
- Create store/useChessAI.ts — React hook that watches turn==='b' and schedules
  the computer move via useRef<timer> + useEffect with cleanup
- Remove let _timer and scheduleComputerMove() from useChessStore.ts
- Remove scheduleComputerMove() call from selectSquare — AI is now triggered
  by the reactive hook, not by the store action
- Mount useChessAI() in ChessGame.tsx so React owns the timer lifecycle
